fix authCheck
