+++ main.go
@@ -3,35 +3,15 @@ package main
 import (
 	"fmt"
 	"log"
-	"net/http"
 	"os"
 
 	_ "github.com/Amakuchisan/QuestionStore/database"
 	"github.com/Amakuchisan/QuestionStore/route"
 	"github.com/joho/godotenv"
-	"github.com/labstack/echo"
 	"github.com/labstack/echo/middleware"
 	"github.com/stretchr/gomniauth"
 	"github.com/stretchr/gomniauth/providers/google"
 )
-
-func authCheckMiddleware() echo.MiddlewareFunc {
-	return func(next echo.HandlerFunc) echo.HandlerFunc {
-		return func(c echo.Context) error {
-			const googleLoginPath = "/auth/login/google"
-			const googleCallbackPath = "/auth/callback/google"
-			_, err := c.Cookie("auth")
-			if err != nil {
-				path := c.Request().URL.Path
-				if !(path == googleLoginPath || path == googleCallbackPath) {
-					return c.Redirect(http.StatusTemporaryRedirect, googleLoginPath)
-				}
-			}
-
-			return next(c)
-		}
-	}
-}
 
 func main() {
 
@@ -46,7 +26,6 @@ func main() {
 	e.Use(middleware.Logger())
 	e.Pre(middleware.RemoveTrailingSlash())
 	e.Static("/static", "static")
-	e.Use(authCheckMiddleware())
 
 	host := os.Getenv("QS_HOST")
 	googleCallbackURL := fmt.Sprintf("http://%s/auth/callback/google", host)
