+++ route/router.go
@@ -3,6 +3,7 @@ package route
 import (
 	"html/template"
 	"io"
+	"net/http"
 
 	"github.com/Amakuchisan/QuestionStore/database"
 	"github.com/Amakuchisan/QuestionStore/handler"
@@ -25,24 +26,39 @@ func Init() *echo.Echo {
 	e := echo.New()
 	e.Debug = true
 
+	g := e.Group("", authCheckMiddleware())
+
 	renderer := &TemplateRenderer{
 		templates: template.Must(template.ParseGlob("templates/*.html")),
 	}
 	e.Renderer = renderer
 
-	e.GET("/", handler.MainPage)
-	e.GET("/questions/form", handler.QuestionFormHandler)
+	g.GET("/", handler.MainPage)
+	g.GET("/questions/form", handler.QuestionFormHandler)
 	e.GET("/auth/login/:provider", handler.LoginHandler)
 
 	userHandler := handler.NewUserHandler(repository.NewUserModel(database.DB))
 	e.GET("/auth/callback/:provider", userHandler.CallbackHandler)
-	e.GET("/users", userHandler.UserAll)
+	g.GET("/users", userHandler.UserAll)
 	// e.GET("/users/:id", userHandler.DetailUser)
 	// e.DELETE("/users/:id", userHandler.DeleteUser)
 	questionHandler := handler.NewQuestionHandler(repository.NewQuestionModel(database.DB))
-	e.POST("/questions", questionHandler.PostQuestion)
-	e.GET("/questions", questionHandler.QuestionsTitleList)
-	e.GET("/questions/:id", questionHandler.QuestionDetail)
+	g.POST("/questions", questionHandler.PostQuestion)
+	g.GET("/questions", questionHandler.QuestionsTitleList)
+	g.GET("/questions/:id", questionHandler.QuestionDetail)
 
 	return e
+}
+
+func authCheckMiddleware() echo.MiddlewareFunc {
+	return func(next echo.HandlerFunc) echo.HandlerFunc {
+		return func(c echo.Context) error {
+			_, err := c.Cookie("auth")
+			if err != nil {
+				return c.Redirect(http.StatusTemporaryRedirect, "/auth/login/google")
+			}
+
+			return next(c)
+		}
+	}
 }
